trusted-types/TrustedScript.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<script>
  test(t => {
    var script = TrustedScript.escape(STRINGS.unescapedHTML);

    assert_equals(script.toString(), STRINGS.escapedHTML);
  }, "Basic escaping.");

  test(t => {
    var script = TrustedScript.unsafelyCreate(STRINGS.unescapedHTML);

    assert_equals(script.toString(), STRINGS.unescapedHTML);
  }, "Basic unsafe construction.");
</script>
